GetRandomJokeAsync method checks if there are any jokes in the cache (stored in the _jokes list). If there are jokes in the cache, it returns a random joke from the cache. If there are no jokes in the cache, it makes a request to the API and retrieves a new joke. The new joke is then added to the cache and returned to the caller.
By using this caching approach, you can reduce the number of API requests that are made and improve the performance of the application.

ChuckNorrisJoke class is used to represent a joke returned by the Chuck Norris joke API. It has a single property, Value, which represents the text of the joke.

The JsonPropertyName attribute is used to specify that the Value property maps to the value field in the JSON response from the API.